+++ hahomematic/central/__init__.py
@@ -26,7 +26,6 @@ from hahomematic.caches.dynamic import CentralDataCache, DeviceDetailsCache
 from hahomematic.caches.persistent import DeviceDescriptionCache, ParamsetDescriptionCache
 from hahomematic.caches.visibility import ParameterVisibilityCache
 from hahomematic.central import xml_rpc_server as xmlrpc
-from hahomematic.central.command_queue import CommandQueueHandler
 from hahomematic.central.decorators import callback_backend_system, callback_event
 from hahomematic.client.json_rpc import JsonRpcAioHttpClient
 from hahomematic.client.xml_rpc import XmlRpcProxy
@@ -35,6 +34,7 @@ from hahomematic.const import (
     DATETIME_FORMAT_MILLIS,
     DEFAULT_INCLUDE_INTERNAL_PROGRAMS,
     DEFAULT_INCLUDE_INTERNAL_SYSVARS,
+    DEFAULT_MAX_READ_WORKERS,
     DEFAULT_PROGRAM_SCAN_ENABLED,
     DEFAULT_SYSVAR_SCAN_ENABLED,
     DEFAULT_TLS,
@@ -155,7 +155,6 @@ class CentralUnit(PayloadMixin):
 
         CENTRAL_INSTANCES[self.name] = self
         self._connection_checker: Final = ConnectionChecker(central=self)
-        self._command_queue_handler: Final = CommandQueueHandler()
         self._hub: Hub = Hub(central=self)
         self._version: str | None = None
         # store last event received datetime by interface
@@ -183,11 +182,6 @@ class CentralUnit(PayloadMixin):
     def clients(self) -> tuple[hmcl.Client, ...]:
         """Return all clients."""
         return tuple(self._clients.values())
-
-    @property
-    def command_queue_handler(self) -> CommandQueueHandler:
-        """Return the que handler for send commands."""
-        return self._command_queue_handler
 
     @property
     def config(self) -> CentralConfig:
@@ -434,8 +428,6 @@ class CentralUnit(PayloadMixin):
         _LOGGER.debug("STOP: Removing instance")
         if self.name in CENTRAL_INSTANCES:
             del CENTRAL_INSTANCES[self.name]
-
-        await self._command_queue_handler.stop()
 
         # wait until tasks are finished
         await self.looper.block_till_done()
@@ -1422,6 +1414,7 @@ class CentralConfig:
         json_port: int | None = None,
         listen_ip_addr: str | None = None,
         listen_port: int | None = None,
+        max_read_workers: int = DEFAULT_MAX_READ_WORKERS,
         un_ignore_list: list[str] | None = None,
         program_scan_enabled: bool = DEFAULT_PROGRAM_SCAN_ENABLED,
         include_internal_programs: bool = DEFAULT_INCLUDE_INTERNAL_PROGRAMS,
@@ -1448,6 +1441,7 @@ class CentralConfig:
         self.json_port: Final = json_port
         self.listen_ip_addr: Final = listen_ip_addr
         self.listen_port: Final = listen_port
+        self.max_read_workers = max_read_workers
         self.un_ignore_list: Final = un_ignore_list
         self.program_scan_enabled: Final = program_scan_enabled
         self.include_internal_programs: Final = include_internal_programs