+++ changelog.md
@@ -1,9 +1,12 @@
-# Version 2024.9.13 (2024-09-29)
+# Version 2024.10.0 (2024-10-03)
 
-- Make UPDATEABLE optional due to homegear support
-- Raise exception on set_value, put_paramset
+- Add config option for max read workers
+- Disable collector for stop events
 - Improve logging when raising exception
 - Log exception at the most outer service
+- Make UPDATEABLE optional due to homegear support
+- Raise exception on set_value, put_paramset
+- Remove command queue
 
 # Version 2024.9.12 (2024-09-26)